pyproject.toml
[project]
name = "video-devour"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "camel-ai>=0.2.75",
    "funasr>=1.2.7",
    "moviepy>=2.2.1",
    "sentence-transformers==5.1.0",
    "numpy>=1.24.0,<2.3.0",
    "opencv-contrib-python>=4.8.0",
    "opencv-python>=4.8.0",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0.0",
    "scikit-image>=0.20.0",
    "soundfile>=0.12.0",
    "torch>=2.0.0",
    "modelscope[audio]>=1.9.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.2.1",
]
